Add basic framework for ECC encoders and decoders
A generic base class for error correction codes, `GenericCode`, is added,
which can be extended by actual implementations of error correction
codes. This `GenericCode` comes with a `GenericEncoder` and
`GenericDecoder`, which should be suitable for most other codes,
although they might not be optimal.

In `matrix_util.py` there are a number of utility functions for
manipulating parity-check matrices. These can be used to generate a
generator matrix from a parity-check matrix.

`IdentityCode` implements the simplest possible code, one which does not
do any error detection or correction. It simply generates an identity
generator matrix and an empty parity-check matrix.

`HammingCode` implements a simple hamming code, this code uses the
utility functions to create the corresponding generator matrix.
`ExtendedHammingCode` implements an extended hamming code, which is
mostly based on the normal hamming code implementation.