Disconnect encoder from decoder in non-formal output
In the case that we are not formally verifiying the encoder and decoder
modules, the modules are no longer connected. This makes it easier to
see what is going on when looking at them in the yosys show output. Also
this means that optimisations cannot remove the part of the encoder or
decoder.