Add HsiaoCode implementation
This implementation will iterate over all possible Hsiao code with a
specific number of data bits. It will search for the variation with the
most balanced rows, which means that the critical path for every bit is
(almost) equal.

This implementation can be improved by using a better optimisation
metric, however this requries a different implementation of the search.