Add construction based implementation for Hsiao code
This implementation of the Hsiao code uses a construction based
algorithm instead of a search algorithm. This means the matrix can be
generated much faster than by searching all possible options.

Right now it seems that the resulting matrix uses slightly more
resources when synthesized, however this is likely not significant.