Implement Dutta and Touba (2007) SEC-DED-DAEC
The Dutta and Touba code is implemented using the Boolector code
framework to do the optimization. This allows the system to find decent
quality codes relatively quickly, without requiring a custom search
algorithm.

Still, this method for finding such a matrix is relatively slow, on the
order of tens of seconds. It might be worth caching the generated
parity-check matrix, especially if a long time is required to find it.