Add simulation test for BasicController
This implements a basic simulation test for the BasicController. The
BasicController is instantiated and connected to a nMigen memory block.
The simulation then drives the request and response ports randomly to
exercise the operation of the controller. A monitoring task keeps track
of the requests and responses that are generated and checks that
everything makes sense.