Add write-back controller implementation "WriteBackController"
This controller implementation will automatically insert a write-back
request when a read response contains a correctable error. This means
that the error will be automatically fixed in the next cycle after it
was detected. Doing this will somewhat slow down the actual requests,
however this is proportional to the number of errors that were detected.